// libs/enums/hr/role.enum.ts
export enum Role{
  ADMIN = "ADMIN",
  CHIEF_ACCOUNTANT = "CHIEF_ACCOUNTANT",
  ACCOUNTANT_CASH_FUND = "ACCOUNTANT_CASH_FUND",
  ACCOUNTANT_MODERATION = "ACCOUNTANT_MODERATION",
  SALESMAN = "SALESMAN",
  SALESMAN_EGG = "SALESMAN_EGG",
  IMPORTER_EXPORTER = "IMPORTER_EXPORTER",
  CONSTRUCTION_DEPARTMENT = "CONSTRUCTION_DEPARTMENT",
  HUMAN_RESOURCE = "HUMAN_RESOURCE",
  CAMP_ACCOUNTING = "CAMP_ACCOUNTING",
  CAMP_MANAGEMENT = "CAMP_MANAGEMENT",
  CAMP_DIRECTOR = "CAMP_DIRECTOR",
  HATCHERY_ACCOUNTING = "HATCHERY_ACCOUNTING",
  HATCHERY_MANAGEMENT = "HATCHERY_DIRECTOR",
  NONE = "NONE",
}